cleanup - removed duplication, unused scripts, rename unnecessary prefixes
