+++ python/HEPCNN/model_default.py
