+++ python/HEPCNN/model_largekernel.py
